Add algorithm to check if a player is blocked
The algorithm verify for all remanings pieces, and their variants, if
they can be placed on the map for the corresponding player. If no space
is found, then the player is blocked and can't play anymore (his turn is
skipped).

The game can now verify if the game is finished (if all players except
one is blocked, then the game has finish and the remaning player is the
winner).
But there are still things to correct/do for the next commit :
- If a player has no more piece then the game considers him blocked
instead of seeing him as a winner
- The UI remains to be done
- For the moment we can't start a new game without quitting the app
- For the moment we can't choose the number of player